refactor(perl-token): split token facade into SRP modules (#8197)
